refactor!: rename namespace
BREAKINGCHANGE

Replace the namespace MakingSense by Doppler
