+++ Doppler.FeatureToggle.TestConsole/Doppler.FeatureToggle.TestConsole.csproj
@@ -2,8 +2,8 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFrameworks>net46;net6.0;net7.0</TargetFrameworks>
-    <RootNamespace>MakingSense.DopplerFeatureToggle.TestConsole</RootNamespace>
-    <AssemblyName>MakingSense.DopplerFeatureToggle.TestConsole</AssemblyName>
+    <RootNamespace>Doppler.FeatureToggle.TestConsole</RootNamespace>
+    <AssemblyName>Doppler.FeatureToggle.TestConsole</AssemblyName>
   </PropertyGroup>
   <ItemGroup>
     <None Remove="BenchmarkDotNet.Artifacts\**" />
