+++ Doppler.FeatureToggle.Tests/FeatureToggleAccessorBuildingTests.cs
@@ -23,8 +23,8 @@
 // OTHER DEALINGS IN THE SOFTWARE.
 #endregion
 
-using MakingSense.DopplerFeatureToggle.Internal;
-using MakingSense.DopplerFeatureToggle.Tests;
+using Doppler.FeatureToggle.Internal;
+using Doppler.FeatureToggle.Tests;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -35,11 +35,11 @@ using System.Threading;
 #if DNXCORE50
 using Xunit;
 using Test = Xunit.FactAttribute;
-using Assert = MakingSense.DopplerFeatureToggle.Tests.XUnitAssert;
+using Assert = Doppler.FeatureToggle.Tests.XUnitAssert;
 #else
 using NUnit.Framework;
 #endif
-namespace MakingSense.DopplerFeatureToggle.Tests
+namespace Doppler.FeatureToggle.Tests
 {
     [TestFixture]
     public class FeatureToggleAccessorBuildingTests : TestFixtureBase
