+++ Doppler.FeatureToggle/Internal/HttpClient_netstandard1.0.cs
@@ -7,7 +7,7 @@ using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MakingSense.DopplerFeatureToggle.Internal
+namespace Doppler.FeatureToggle.Internal
 {
     /// <summary>
     /// Custom HttpClient implementation
